requirements.lock.txt
torch>=2.6,<2.7
transformers>=4.49,<4.51
huggingface-hub>=0.30,<1.0
PyYAML>=6.0,<7.0
safetensors>=0.5,<1.0
